add alias Faraday::FilePart for older Faraday::UploadIO
