+++ lib/faraday/param_part.rb
@@ -1,18 +1,34 @@
 # frozen_string_literal: true
 
 module Faraday
-  # Allows multipart posts while specifying headers of a part
   class ParamPart
+    # Multipart value used to POST data with a content type.
+    #
+    # @param value [String] Uploaded content as a String.
+    # @param content_type [String] String content type of the value.
+    # @param content_id [String] Optional String of this value's Content-ID.
+    #
+    # @return [Faraday::ParamPart]
     def initialize(value, content_type, content_id = nil)
       @value = value
       @content_type = content_type
       @content_id = content_id
     end
 
+    # Converts this value to a form part.
+    #
+    # @param boundary [String] String multipart boundary that must not exist in
+    #   the content exactly.
+    # @param key [String] String key name for this value.
+    #
+    # @return [Faraday::Parts::Part]
     def to_part(boundary, key)
       Faraday::Parts::Part.new(boundary, key, value, headers)
     end
 
+    # Returns a Hash of String key/value pairs.
+    #
+    # @return [Hash]
     def headers
       {
         'Content-Type' => content_type,
@@ -20,6 +36,19 @@ module Faraday
       }
     end
 
-    attr_reader :value, :content_type, :content_id
+    # The content to upload.
+    #
+    # @return [String]
+    attr_reader :value
+
+    # The value's content type.
+    #
+    # @return [String]
+    attr_reader :content_type
+
+    # The value's content ID, if given.
+    #
+    # @return [String, nil]
+    attr_reader :content_id
   end
 end
